break some comps into separate files. move session operations to module
